Moved the supporting python files to the lib directory per kodi specs
