+++ resources/lib/adobe_activate_api.py
