+++ resources/lib/util.py
